Don't add test dir when it doesn't exist
Closes #1

The `spec` dir is treated special, becaues it's a default dir.
In our case, we want to add a second default dir, but that's not a thing.

I was trying to mirror this code:
  https://github.com/rspec/rspec-core/blob/5375dcdfdb5460a51d5003de76eb124997fa26ea/lib/rspec/core/configuration.rb#L835

But it does not get past this statement:
  https://github.com/rspec/rspec-core/blob/5375dcdfdb5460a51d5003de76eb124997fa26ea/lib/rspec/core/configuration.rb#L1725

So `spec` will be filtered out if it DNE, but `test` will not
  https://github.com/rspec/rspec-core/blob/5375dcdfdb5460a51d5003de76eb124997fa26ea/lib/rspec/core/configuration.rb#L1659

And it will eventually try to load "test", thinking it's a file, causing it to blow up
  https://github.com/rspec/rspec-core/blob/5375dcdfdb5460a51d5003de76eb124997fa26ea/lib/rspec/core/configuration.rb#L1320